refactor(coding-agent): keep thinking validation local
