+++ packages/coding-agent/src/cli/args.ts
@@ -4,7 +4,7 @@
 
 import type { ThinkingLevel } from "@earendil-works/pi-agent-core";
 import { APP_NAME } from "../config.js";
-import { isValidThinkingLevel, THINKING_LEVELS } from "../core/thinking-levels.js";
+import { THINKING_LEVELS } from "../core/thinking-levels.js";
 
 export type Mode = "text" | "json" | "rpc" | "acp" | "daemon";
 
@@ -64,6 +64,10 @@ const REMOVED_BUILTIN_TOOL_NAMES = new Set(["read", "write", "grep", "find", "ls
 const BUILTIN_TOOL_NAMES = ["ipython"];
 
 export const INTERNAL_RUNTIME_COMMAND_MARKER = "\0prime-agent-runtime-command";
+
+export function isValidThinkingLevel(level: string): level is ThinkingLevel {
+	return THINKING_LEVELS.includes(level as ThinkingLevel);
+}
 
 export function parseArgs(args: string[]): Args {
 	const result: Args = {
